지수 띄우기 구현
