+++ zeroback/src/component/IndexPage/IndexPage.css
@@ -1,0 +1,6 @@
+.index-box {
+    margin-top: 30px;
+    margin-bottom: 50px;
+    display: flex;
+    justify-content: center;
+}
