+++ zeroback/src/index.js
@@ -7,9 +7,7 @@ import reportWebVitals from './reportWebVitals';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <React.StrictMode>
-    <BrowserRouter><App /></BrowserRouter>
-  </React.StrictMode>
+  <BrowserRouter><App /></BrowserRouter>
 );
 
 // If you want to start measuring performance in your app, pass a function
